+++ src/utils/parseError.ts
@@ -1,4 +1,4 @@
-import { Process, Step } from '@lifi/types'
+import { LifiStep, Process } from '@lifi/types'
 import {
   errorCodes as MetaMaskErrorCodes,
   getMessageFromCode,
@@ -56,7 +56,7 @@ import { TenderlyResponse } from '..'
  */
 
 export const getTransactionNotSentMessage = async (
-  step?: Step,
+  step?: LifiStep,
   process?: Process
 ): Promise<string> => {
   let transactionNotSend =
@@ -86,7 +86,7 @@ export const getTransactionNotSentMessage = async (
 }
 
 export const getTransactionFailedMessage = async (
-  step: Step,
+  step: LifiStep,
   txLink?: string
 ): Promise<string> => {
   const chainsService = ChainsService.getInstance()
@@ -102,7 +102,7 @@ export const getTransactionFailedMessage = async (
 
 export const parseError = async (
   e: any,
-  step?: Step,
+  step?: LifiStep,
   process?: Process
 ): Promise<LifiError> => {
   if (e instanceof LifiError) {